Fixed syntax error

+++ process/processor.go
@@ -478,7 +478,7 @@ func processRevealEntry(msg *wire.MsgRevealEntry) error {
 		}
 
 		// 37 effectively removes the entry header.  1023 rounds up the credit calucation.
-		cred := (len(bin) - 37 + 1023) / 1024
+		cred := int32((len(bin) - 37 + 1023) / 1024)
 
 		if int32(c.Credits) < cred {
 			fMemPool.addOrphanMsg(msg, h)
